sécuritéBack - roleAdmin

+++ controllers/suppr-article-controller.php
@@ -4,6 +4,8 @@ require_once DOSSIER_MODEL . '/Article.php';
 $article = Article::retrieveByPK($_GET['id']);
 
 if (empty($article)) error();
-$article->delete();
 
-redirection("liste-articles");
+if ($_SESSION['role'] == "admin") {
+    $article->delete();
+    redirection("liste-articles");
+} else redirection("liste-articles");
